Extra FailureStrategy to own class
This is a purely mechanical change
