+++ server/src/main/java/org/elasticsearch/indices/cluster/IndicesClusterStateService.java
@@ -75,6 +75,7 @@ import org.elasticsearch.index.shard.ShardId;
 import org.elasticsearch.index.shard.ShardLongFieldRange;
 import org.elasticsearch.index.shard.ShardNotFoundException;
 import org.elasticsearch.indices.IndicesService;
+import org.elasticsearch.indices.recovery.FailureStrategy;
 import org.elasticsearch.indices.recovery.PeerRecoverySourceService;
 import org.elasticsearch.indices.recovery.PeerRecoveryTargetService;
 import org.elasticsearch.indices.recovery.RecoveryCancelledException;
@@ -1311,7 +1312,7 @@ public class IndicesClusterStateService extends AbstractLifecycleComponent imple
     // package-private for testing
     synchronized void handleRecoveryFailure(
         ShardRouting shardRouting,
-        RecoveryListener.FailureStrategy failureStrategy,
+        FailureStrategy failureStrategy,
         long primaryTerm,
         Exception failure
     ) {
